menu, improved makefile
