+++ learn-silice/classroom/soc_wave_player/firmware/Makefile
@@ -3,15 +3,19 @@ CC   = $(ARCH)-gcc
 AS   = $(ARCH)-as
 LD   = $(ARCH)-ld
 
-main : main.o div.o crt0.o fat_io_lib.o
+test_fat32 : test_fat32.o div.o crt0.o fat_io_lib.o
 		$(LD) -m elf32lriscv -b elf32-littleriscv -Tconfig_c.ld --no-relax -o code.elf $^
 		$(ARCH)-objcopy -O verilog code.elf code.hex
 
-test_sdcard : test_sdcard.o div.o crt0.o fat_io_lib.o
+test_sdcard : test_sdcard.o div.o crt0.o
 		$(LD) -m elf32lriscv -b elf32-littleriscv -Tconfig_c.ld --no-relax -o code.elf $^
 		$(ARCH)-objcopy -O verilog code.elf code.hex
 
-test_hello_world: test_hello_world.o div.o crt0.o fat_io_lib.o
+test_hello_world: test_hello_world.o div.o crt0.o
+		$(LD) -m elf32lriscv -b elf32-littleriscv -Tconfig_c.ld --no-relax -o code.elf $^
+		$(ARCH)-objcopy -O verilog code.elf code.hex
+
+test_menu: test_menu.o div.o crt0.o
 		$(LD) -m elf32lriscv -b elf32-littleriscv -Tconfig_c.ld --no-relax -o code.elf $^
 		$(ARCH)-objcopy -O verilog code.elf code.hex
