+++ learn-silice/classroom/soc_wave_player/firmware/test_fat32.c
@@ -1,20 +1,7 @@
 // @sylefeb 2022-01-10
 
 #include "config.h"
-
-static inline unsigned int rdcycle()
-{
-   unsigned int cycles;
-   asm volatile ("rdcycle %0" : "=r"(cycles));
-   return cycles;
-}
-
-static inline void pause(int ncycles)
-{
-  unsigned int start = rdcycle();
-  while ( rdcycle() - start < ncycles ) { }
-}
-
+#include "std.c"
 #include "oled.h"
 #include "printf.c"
 #include "mul.c"
@@ -22,19 +9,6 @@ static inline void pause(int ncycles)
 #include "display.c"
 
 #include "fat_io_lib/src/fat_filelib.h"
-
-void uart_putchar(int c)
-{
-  *UART = c;
-  pause(10000);
-}
-
-void dual_putchar(int c)
-{
-  display_putchar(c);
-  *UART = c;
-  pause(10000);
-}
 
 void main()
 {
